# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post API",
        "description": "API to manage posts",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts",
                "description": "Retrieve a list of all posts",
                "operationId": "baefad547d39da5fb3215d82afeb6a9e",
                "responses": {
                    "200": {
                        "description": "List of posts"
                    },
                    "404": {
                        "description": "No student found"
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "description": "Create a new student with provided name and age",
                "operationId": "9db5f611b7148c9346ac6203619abb37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get a post by ID",
                "description": "Retrieve a post by its ID",
                "operationId": "1b5891692e7747118874a4e5903004d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "post found"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "posts"
                ],
                "summary": "Update a student",
                "description": "Update the details of a post",
                "operationId": "d1e4f0a694e9d7378f3dd5599fc0ad12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the student to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "post updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "description": "Delete a post by its ID",
                "operationId": "cc044a483f4d291da4c6e4875bcf0509",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "post deleted"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            }
        }
    }
}